requirements.txt
asgiref==3.2.7
Django==3.0.4
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
django-extensions==2.2.9
ipython==7.13.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
